Create ID of interest
This file contains the ids of animals to be linked to the pedigree. The file has two columns. Column one has the id of the animals that corresponds to the id column in the pedigree file. Column two has the group of the animals, which is used to determine the activeness of the animal per group or population.